moved docx code out from main to its own module
